+++ react/components/SearchQuery.js
@@ -7,6 +7,7 @@ import { onSearchResult } from "vtex.sae-analytics";
 import BiggyClient from "../utils/biggy-client.ts";
 import {
   makeFetchMore,
+  makeRefetch,
   fromAttributesToFacets,
 } from "../utils/compatibility-layer.ts";
 import logError from "../utils/log.ts";
@@ -35,11 +36,12 @@ const SearchQuery = ({
     variables,
     ssr: false,
     fetchPolicy: "network-only",
-    onCompleted: data => {
-      saveTermInHistory(variables.query);
-      onSearchResult(data);
-    },
   });
+
+  useEffect(() => saveTermInHistory(variables.query), [variables.query]);
+  useEffect(() => searchResult.data && onSearchResult(searchResult.data), [
+    searchResult.data,
+  ]);
 
   if (searchResult.error) {
     logError(account, workspace, attributePath, searchResult.error);
@@ -51,6 +53,7 @@ const SearchQuery = ({
   const products = path(["data", "searchResult", "products"], searchResult);
 
   const fetchMore = makeFetchMore(searchResult.fetchMore, page, setPage);
+  const refetch = makeRefetch(searchResult.refetch);
   const recordsFiltered = pathOr(
     0,
     ["data", "searchResult", "total"],
@@ -103,7 +106,7 @@ const SearchQuery = ({
 
   searchQuery.loading = searchResult.loading;
   searchQuery.fetchMore = fetchMore;
-  searchQuery.refetch = () => searchResult.refetch();
+  searchQuery.refetch = refetch;
 
   return children({
     ...searchQuery.variables,